Improves web syndication page design
